backend/requirements.txt
# ================================
# BACKEND DEPENDENCIES
# ================================

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching & Sessions
redis==5.0.1
celery==5.3.4

# AI and Agent Libraries
crewai==0.28.8
cerebras-cloud-sdk==1.3.0
crewai-tools==0.4.26
langchain==0.1.0
langchain-community==0.0.10

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0

# Monitoring
psutil==5.9.6
sentry-sdk[fastapi]==1.38.0

# SSO (Optional)
authlib==1.2.1 